Implemented the datasource manager to get the tenant database.
I needed to change the loading cache to be based on the tenant. In turn, since
the tenant was used in a set (the cache) needed to ensure it was unique by the
id. (Maybe the tenant uuid is useless? May remove). Wrote unit tests for the
lot, while real test against the liquibase code