Updating infrastructure for new projects.
Kinda sizable change. Adding the control and proxy server. Creating a
common server project so resources that are not specific to dropwizard
services can be easily shared. I'd like to do the dagger creationg of
the Server runtime, but that might take a bit. This should at least
reduce the bloat.