Remove SqlEngine, replace with Jdbi
Last bit of SqlEngine replaced with Jdbi.  The structure of the
Tenant_data table is not very friendly to Jdbi, so it's a bit messier
than the other migrations though.

As part of this, the SqlEngine instance was also removed from the
BaseSQLTest, though it appeared to be unused anyway.

This completes issue #4