Added expiry to tenant data for tombstones.
I need to update liquibase so the index is only on entries where
expiry is not null. hsqldb does not support that but looks like
it allows it for compatibility. May need to switch to sqlite
eventually because that kinda sucks. I'll add a task to deep dive
to see the impact there.